requirements.txt
watchdog
gitpython